NFT Series can now be owned. Changed Series Id frm u64 to u32.

+++ pallets/marketplace/src/tests/mock.rs
@@ -75,7 +75,7 @@ impl ternoa_nfts::Config for Test {
     type NFTId = u8;
     type NFTDetails = MockNFTDetails;
     type WeightInfo = ();
-    type NFTSeriesId = u64;
+    type NFTSeriesId = u32;
 }
 
 impl Config for Test {
