+++ pallets/timed-escrow/src/tests/mock.rs
@@ -72,7 +72,7 @@ impl ternoa_nfts::Config for Test {
     type NFTId = u8;
     type NFTDetails = ();
     type WeightInfo = ();
-    type NFTSeriesId = u64;
+    type NFTSeriesId = u32;
 }
 impl Config for Test {
     type Event = Event;
